backend/requirements.txt
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.23.2,<0.24.0
redis>=5.0.0,<6.0.0
PyPDF2>=3.0.0,<4.0.0
google-generativeai>=0.3.0,<0.4.0
python-multipart>=0.0.6,<0.1.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
httpx>=0.25.0,<0.26.0
reportlab>=4.0.0,<5.0.0

# Testing dependencies
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0
fakeredis>=2.20.0,<3.0.0
pytest-cov>=4.1.0,<5.0.0 